[Format]: Replace standard library output with fmt lib.

+++ tests/frontend/basic-types.c
@@ -1,1 +1,1 @@
-char test(char a, int b) { return a + b; }
+char test(char a, double b) { return a + b;}
